new logic to specially handle paddle collisions

+++ src/core/system/types/CollisionSystem.js
@@ -1,7 +1,10 @@
+import PaddleCollisionComponent from "../../../extensions/component/PaddleCollisionComponent";
 import componentManager from "../../component/ComponentManager";
 import BoxCollisionComponent from "../../component/types/BoxCollisionComponent";
 import DestructableComponent from "../../component/types/DestructableComponent";
+import LocationComponent from "../../component/types/LocationComponent";
 import VelocityComponent from "../../component/types/VelocityComponent";
+import { X } from "../../consts";
 import entityManager from "../../entity/EntityManager";
 
 const pointIsWithinBox = (point, collisionComp) => {
@@ -32,6 +35,25 @@ function getRndInteger(min, max) {
   return Math.floor(Math.random() * (max - min)) + min;
 }
 
+const handlePaddleCollision = (paddle, vel, loc) => {
+  if (loc) {
+    const locx = loc.x;
+    if (locx >= paddle.cutPoints[0] && locx <= paddle.cutPoints[1]) {
+      if (vel.vec.vec[X] < 0) {
+        vel.vec.reverseY();
+      } else {
+        vel.vec.reverse();
+      }
+    } else if (locx > paddle.cutPoints[1] && locx <= paddle.cutPoints[2]) {
+      if (vel.vec.vec[X] > 0) {
+        vel.vec.reverseY();
+      } else {
+        vel.vec.reverse();
+      }
+    }
+  }
+};
+
 const handleCollision = (ent1, ent2) => {
   const vel1 = componentManager.lookupComponent(
     VelocityComponent.getComponentId(),
@@ -51,6 +73,38 @@ const handleCollision = (ent1, ent2) => {
     DestructableComponent.getComponentId(),
     ent2
   );
+
+  const paddleCollision1 = componentManager.lookupComponent(
+    PaddleCollisionComponent.getComponentId(),
+    ent1
+  );
+
+  const paddleCollision2 = componentManager.lookupComponent(
+    PaddleCollisionComponent.getComponentId(),
+    ent2
+  );
+
+  if (paddleCollision1) {
+    const ent2Loc = componentManager.lookupComponent(
+      LocationComponent.getComponentId(),
+      ent2
+    );
+
+    handlePaddleCollision(paddleCollision1, vel2, ent2Loc);
+
+    return;
+  }
+
+  if (paddleCollision2) {
+    const ent1Loc = componentManager.lookupComponent(
+      LocationComponent.getComponentId(),
+      ent1
+    );
+
+    handlePaddleCollision(paddleCollision2, vel1, ent1Loc);
+
+    return;
+  }
 
   if (dest1) entityManager.removeEntity(ent1.id);
   if (dest2) entityManager.removeEntity(ent2.id);
@@ -92,11 +146,19 @@ class CollisionSystem {
             componentManager.lookupComponent(
               BoxCollisionComponent.getComponentId(),
               firstEntity
-            ),
+            ) ||
+              componentManager.lookupComponent(
+                PaddleCollisionComponent.getComponentId(),
+                firstEntity
+              ),
             componentManager.lookupComponent(
               BoxCollisionComponent.getComponentId(),
               secondEntity
-            ),
+            ) ||
+              componentManager.lookupComponent(
+                PaddleCollisionComponent.getComponentId(),
+                secondEntity
+              ),
           ];
 
           if (!collisionModels[0] || !collisionModels[1]) continue;
